update pr based on comments
