+++ rust-runtime/aws-smithy-types/src/lib.rs
@@ -21,7 +21,6 @@ pub mod config_bag;
 pub mod date_time;
 pub mod endpoint;
 pub mod error;
-pub mod macros;
 pub mod primitive;
 pub mod retry;
 pub mod timeout;
